requirements.txt
absl-py==0.5.0
astor==0.7.1
bleach==1.5.0
gast==0.2.0
grpcio==1.15.0
h5py==2.8.0
html5lib==0.9999999
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
numpy==1.15.2
opencv-python==3.4.3.18
pandas==0.23.4
protobuf==3.6.1
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
sklearn==0.0
tensorboard==1.8.0
tensorflow-gpu==1.8.0
termcolor==1.1.0
Werkzeug==0.14.1
